prairiedog: add reboot-if-required command
When boot settings change, "prairiedog generate-boot-config" is invoked
to regenerate the initrd containing the boot config. However, this does
not mean that any changes to e.g. the kernel or systemd command lines
take effect yet. A reboot is needed for this.

Add a new "prairiedog reboot-if-required" command that compares the
current boot settings (and hence the currently persisted boot config in
the initrd) against the one the host booted with to detect any changes
requiring a reboot. If such changes are detected and the
"settings.boot.reboot-to-reconcile" flag is set, reboot the host.

For now, only changes to the kernel and systemd command lines require a
reboot. Bottlerocket-specific boot settings like the newly introduced
"reboot-to-reconcile" may change freely and will not be flagged by
prairiedog.

Signed-off-by: Markus Boehme <markubo@amazon.com>