os, release: reboot host if needed after reaching configured.target
Reboot the system towards the end of the configuration phase during boot
if any of the first-party components deem this necessary. The reboot is
enacted after reaching the configured.target and before reaching the
multi-user.target, so e.g. user data has been applied and bootstrap
containers have already run, but the host is not yet fully operational.

Components can hook into the new reboot-if-required.service to be able
to run at this point. They should implement a "reboot-if-required"
command to trigger a reboot if it is needed.

Currently, prairiedog is the only component hook into this service and
implement implement a "reboot-if-required" command. It will trigger a
reboot if boot settings have changed, i.e. if the kernel or systemd
command line arguments were changed via the user data mechanism or a
bootstrap container.

Signed-off-by: Markus Boehme <markubo@amazon.com>