build: ensure empty build/rpms directory is included in build context
Whether intentionally or not, Docker 23 subtly changed the handling of
negated patterns in `.dockerignore` files. In our case,

    /build/*
    !/build/rpms/*.rpm

will not include `build/rpms` in the build context with Docker 23 if
the `build/rpms` is empty. This breaks the build since the directory is
expected to exist in any case.

Add a couple extra patterns that amount to the desired behavior, whether
run with Docker 23 or prior versions.

Signed-off-by: Markus Boehme <markubo@amazon.com>
(cherry picked from commit 1e3271751733176994fd02874f650191ff19152a)