HACK: build: make srpm available during build
May not actually be needed. Previous attempt of wiring up kpatch-build used
the SRPM instead of the local source dir. If we keep live patches with the
kernel, however, we can just use the local source dir and drop this patch
again.

While `rpmbuild -bb` should build both source and binary PRMs--and actually
does so--the SRPM is not available until after build, which throws a wrench
into our plans of using it for building live patches.

Signed-off-by: Markus Boehme <markubo@amazon.com>